Remove unused scratch buffers and clean up tests
