+++ xml-json-spring-boot-starter/test/java/com/example/transformer/LargeStreamTest.java
@@ -50,41 +50,4 @@ public class LargeStreamTest {
         assertTrue(diff < 50 * 1024 * 1024);
     }
 
-    @Test
-    public void streamerLargeDocument() throws Exception {
-        XmlToJsonStreamer streamer = XmlToJsonStreamer.builder()
-                .mappingConfig(new MappingConfig())
-                .build();
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("<items>");
-        for (int i = 0; i < 100000; i++) {
-            sb.append("<item>").append(i).append("</item>");
-        }
-        sb.append("</items>");
-        byte[] xml = sb.toString().getBytes(StandardCharsets.UTF_8);
-
-        MemoryMXBean bean = ManagementFactory.getMemoryMXBean();
-        MemoryUsage before = bean.getHeapMemoryUsage();
-        ByteArrayOutputStream out = new ByteArrayOutputStream();
-        streamer.transform(new ByteArrayInputStream(xml), out);
-        MemoryUsage after = bean.getHeapMemoryUsage();
-        long diff = after.getUsed() - before.getUsed();
-        assertTrue(diff < 50 * 1024 * 1024);
-
-        String json = out.toString(StandardCharsets.UTF_8);
-        JsonFactory factory = new JsonFactory();
-        JsonParser p = factory.createParser(json);
-        assertEquals(JsonToken.START_OBJECT, p.nextToken());
-        assertEquals("items", p.nextFieldName());
-        assertEquals(JsonToken.START_OBJECT, p.nextToken());
-        assertEquals("item", p.nextFieldName());
-        assertEquals(JsonToken.START_ARRAY, p.nextToken());
-        int count = 0;
-        while (p.nextToken() != JsonToken.END_ARRAY) {
-            p.getValueAsString();
-            count++;
-        }
-        assertEquals(100000, count);
-    }
 }
